+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## WIP
 
+## 0.4.0
+
 - Reduce generic instantiations internally produced by `ExactLengthArrayPool<T>.Shared` and `ExactLengthArrayPool<T>.SharedOfLength(int)`.
 - Fix error in `ExactLengthArrayPool<T>.Shared.ApproximateCount()` and `ExactLengthArrayPool<T>.Shared.Trim(bool)`.
 - Fix `ExactLengthArrayPool<T>.Shared` and `ExactLengthArrayPool<T>.SharedOfLength(int)` returned incorrect array lengths.
@@ -9,6 +11,17 @@
 - Avoid contention when calling `FastObjectPool<T>.Reserve` and `SafeObjectPool<T>.Reserve`.
 - Replace `IsReserveDynamic` with `IsReserveFixed` in `FastObjectPool<T>`, `SafeObjectPool<T>`, `SafeExactLengthArrayObjectPool<T>` and `SafeExactLengthArrayPool<T>`.
 - Fix `SafeExactLengthArrayPool<T>` creating non-configured instances of `SafeExactLengthArrayObjectPool<T>`.
+- Improve trimming support by correctly annotating and modifying APIs:
+  - Remove nullability in `new FastObjectPool<T>(Func<T>?)` and `new SafeObjectPool<T>(Func<T>?)`.
+  - Remove parameterless constructor `new FastObjectPool<T>()` and `new FastObjectPool<T>()`.
+  - Remove `ObjectPool<T>.Shared`.
+  - Add `ObjectPool` class, which includes extension members for `ObjectPool<T>` and is correctly annotated to require dynamic access to the public parameterless constructor:
+    - `ObjectPool<T>.Shared`.
+    - `ObjectPool<T>.Factory`.
+    - `ObjectPool<T>.CreateInstance()`.
+  - Add `FastObjectPool` and `SafeObjectPool` classes, which includes extension members for `FastObjectPool<T>` and `SafeObjectPool<T>`, and are correctly annotated to require dynamic access to the public parameterless constructor:
+    - `FastObjectPool<T>.CreateDefault()`.
+    - `SafeObjectPool<T>.CreateDefault()`.
 
 ## v0.3.1